+++ src/frontend/configuration/wording/WordingFr.tsx
@@ -209,6 +209,7 @@ export class WordingFr implements Wording {
   readonly TAUX_OCCUPATION_INTERNAT: string = "Taux d’occupation internat ";
   readonly TAUX_OCCUPATION_AUTRE: string = "Taux d'occupation Autre 1, 2 et 3";
   readonly TAUX_OCCUPATION_SEANCES: string = "Taux d'occupation séances";
+  readonly TAUX_OCCUPATION_GLOBAL: string = "Taux d'occupation global";
 
   // Bloc Activité Sanitaire
   readonly NOMBRE_DE_SÉJOUR_MCO: string = "Nombre de séjours Médecine, Chirurgie et Obstétrique en Hospitalisation Complète et Hospitalisation Partielle";
@@ -636,6 +637,7 @@ export class WordingFr implements Wording {
   readonly CREATION_DATE: string = "Date de création";
   readonly PROFILE_CODE: string = "Code";
   readonly PROFILE_TITLE: string = "Libellé de l'autorisation";
+  readonly NEW_PROFILE_LABEL_MANDATORY: string = "Veuillez renseigner un libellé.";
   readonly VOUS_NAVEZ_AUCUN_PROFIL: string = "Vous n'avez aucun profil";
   readonly PROJECT_TEAM: string = "Équipe projet";