+++ src/frontend/ui/établissement-territorial-médico-social/bloc-activité/PageÉtablissementTerritorialMédicoSocialActivité.test.tsx
@@ -64,7 +64,6 @@ describe("La page établissement territorial médico-social Sauf EHPAD - bloc ac
     // [wording.TAUX_OCCUPATION_HÉBERGEMENT_PERMANENT, 0, "CNSA", wording.CNSA_TITLE],
     // [wording.TAUX_OCCUPATION_HÉBERGEMENT_TEMPORAIRE, 1, "CNSA", wording.CNSA_TITLE],
     // [wording.TAUX_OCCUPATION_ACCUEIL_DE_JOUR, 2, "CNSA", wording.CNSA_TITLE],
-    [wording.TAUX_RÉALISATION_ACTIVITÉ, 5, "TdB Perf", wording.TDB_PERF_TITLE],
     [wording.FILE_ACTIVE_PERSONNES_ACCOMPAGNÉES, 6, "TdB Perf", wording.TDB_PERF_TITLE],
     [wording.NOMBRE_MOYEN_JOURNÉES_ABSENCE_PERSONNES_ACCOMPAGNÉES, 7, "TdB Perf", wording.TDB_PERF_TITLE],
     [wording.DURÉE_MOYENNE_SÉJOUR_ACCOMPAGNEMENT_PERSONNES_SORTIES, 8, "TdB Perf", wording.TDB_PERF_TITLE],
@@ -73,6 +72,7 @@ describe("La page établissement territorial médico-social Sauf EHPAD - bloc ac
     // [wording.TAUX_OCCUPATION_INTERNAT, 6, "CNSA", wording.CNSA_TITLE],
     [wording.TAUX_OCCUPATION_AUTRE, 3, "CNSA", wording.CNSA_TITLE],
     [wording.TAUX_OCCUPATION_SEANCES, 4, "CNSA", wording.CNSA_TITLE],
+    [wording.TAUX_OCCUPATION_GLOBAL, 5, "CNSA", wording.CNSA_TITLE],
   ])("affiche les informations l’indicateur %s", (titreSection, identifiant, sourceOrigineAttendue, abréviationSourceOrigineAttendue) => {
     // WHEN
     renderFakeComponent(<SessionProvider session={mockSession}><PageÉtablissementTerritorialMédicoSocial rechercheViewModel={rechercheViewModel} établissementTerritorialViewModel={établissementTerritorialMédicoSocial} /></SessionProvider>);
@@ -100,7 +100,6 @@ describe("La page établissement territorial médico-social Sauf EHPAD - bloc ac
   });
 
   it.each([
-    [wording.TAUX_RÉALISATION_ACTIVITÉ, 5, "100,4 %", "94,5 %", "96,6 %"],
     [wording.FILE_ACTIVE_PERSONNES_ACCOMPAGNÉES, 6, "340", "280", "300"],
     [wording.NOMBRE_MOYEN_JOURNÉES_ABSENCE_PERSONNES_ACCOMPAGNÉES, 7, "87", "90", "22"],
     [wording.DURÉE_MOYENNE_SÉJOUR_ACCOMPAGNEMENT_PERSONNES_SORTIES, 8, (1013).toLocaleString("fr"), "994", "990"],
@@ -201,6 +200,10 @@ describe("La page établissement territorial médico-social Sauf EHPAD - bloc ac
               tauxOccupationSeances: {
                 value: valeurIndicateur1,
               },
+              tauxOccupationGlobal: {
+                dateMiseÀJourSource: "2021-07-07",
+                value: valeurIndicateur1,
+              },
             },
             {
               année: 2021,
@@ -246,6 +249,10 @@ describe("La page établissement territorial médico-social Sauf EHPAD - bloc ac
                 value: valeurIndicateur2,
               },
               tauxOccupationSeances: {
+                value: valeurIndicateur2,
+              },
+              tauxOccupationGlobal: {
+                dateMiseÀJourSource: "2021-07-07",
                 value: valeurIndicateur2,
               },
             },
@@ -295,6 +302,10 @@ describe("La page établissement territorial médico-social Sauf EHPAD - bloc ac
               tauxOccupationSeances: {
                 value: valeurIndicateur3,
               },
+              tauxOccupationGlobal: {
+                dateMiseÀJourSource: "2021-07-07",
+                value: valeurIndicateur3,
+              },
             },
           ],
           autorisationsEtCapacités: ÉtablissementTerritorialMédicoSocialViewModelTestBuilder.autorisations,
@@ -315,7 +326,7 @@ describe("La page établissement territorial médico-social Sauf EHPAD - bloc ac
             motifsRuptureContratLibelles: [],
             natureContratsAnnuel: [],
             natureContratsTrimestriel: [],
-            echelleTemporelle:{}
+            echelleTemporelle: {}
           },
           qualite: { reclamations: [], evenementsIndesirables: [], inspectionsEtControles: { dateMiseAJourSource: '202-02-02', inspectionsEtControles: [] } },
           autorisations: autorisationsMockData,
@@ -402,6 +413,10 @@ describe("La page établissement territorial médico-social Sauf EHPAD - bloc ac
               tauxOccupationSeances: {
                 value: valeurIndicateur1,
               },
+              tauxOccupationGlobal: {
+                dateMiseÀJourSource: "2021-07-07",
+                value: valeurIndicateur1,
+              },
             },
             {
               année: 2021,
@@ -447,6 +462,10 @@ describe("La page établissement territorial médico-social Sauf EHPAD - bloc ac
                 value: valeurIndicateur2,
               },
               tauxOccupationSeances: {
+                value: valeurIndicateur2,
+              },
+              tauxOccupationGlobal: {
+                dateMiseÀJourSource: "2021-07-07",
                 value: valeurIndicateur2,
               },
             },
@@ -496,6 +515,10 @@ describe("La page établissement territorial médico-social Sauf EHPAD - bloc ac
               tauxOccupationSeances: {
                 value: valeurIndicateur3,
               },
+              tauxOccupationGlobal: {
+                dateMiseÀJourSource: "2021-07-07",
+                value: valeurIndicateur3,
+              },
             },
           ],
           autorisationsEtCapacités: ÉtablissementTerritorialMédicoSocialViewModelTestBuilder.autorisations,
@@ -541,7 +564,6 @@ describe("La page établissement territorial médico-social Sauf EHPAD - bloc ac
   );
 
   it.each([
-    [wording.TAUX_RÉALISATION_ACTIVITÉ, 5, "TdB Perf", wording.TDB_PERF_TITLE],
     [wording.FILE_ACTIVE_PERSONNES_ACCOMPAGNÉES, 6, "TdB Perf", wording.TDB_PERF_TITLE],
     [wording.NOMBRE_MOYEN_JOURNÉES_ABSENCE_PERSONNES_ACCOMPAGNÉES, 7, "TdB Perf", wording.TDB_PERF_TITLE],
     [wording.DURÉE_MOYENNE_SÉJOUR_ACCOMPAGNEMENT_PERSONNES_SORTIES, 8, "TdB Perf", wording.TDB_PERF_TITLE],
@@ -550,6 +572,7 @@ describe("La page établissement territorial médico-social Sauf EHPAD - bloc ac
     // [wording.TAUX_OCCUPATION_INTERNAT, 6, "CNSA", wording.CNSA_TITLE],
     [wording.TAUX_OCCUPATION_AUTRE, 3, "CNSA", wording.CNSA_TITLE],
     [wording.TAUX_OCCUPATION_SEANCES, 4, "CNSA", wording.CNSA_TITLE],
+    [wording.TAUX_OCCUPATION_GLOBAL, 5, "CNSA", wording.CNSA_TITLE],
   ])(
     'affiche le contenu de l’info bulle après avoir cliqué sur le bouton "détails" (%s)',
     (titreSection, identifiant, sourceOrigineAttendue, abréviationSourceOrigineAttendue) => {
@@ -583,15 +606,15 @@ describe("La page établissement territorial médico-social Sauf EHPAD - bloc ac
   );
 
   it.each([
-    [wording.TAUX_RÉALISATION_ACTIVITÉ, 5],
-    // [wording.FILE_ACTIVE_PERSONNES_ACCOMPAGNÉES, 6],
+    //[wording.TAUX_RÉALISATION_ACTIVITÉ, 5],
+    // [wording.FILE_ACTIVE_PERSONNES_ACCOMPAGNÉES, 4],
     // [wording.NOMBRE_MOYEN_JOURNÉES_ABSENCE_PERSONNES_ACCOMPAGNÉES, 7],
     // [wording.DURÉE_MOYENNE_SÉJOUR_ACCOMPAGNEMENT_PERSONNES_SORTIES, 8],
-    // // [wording.TAUX_OCCUPATION_EXTERNAT, 4],
+    //[wording.TAUX_OCCUPATION_EXTERNAT, 4],
     // // [wording.TAUX_OCCUPATION_SEMI_INTERNAT, 5],
     // // [wording.TAUX_OCCUPATION_INTERNAT, 6],
     // [wording.TAUX_OCCUPATION_AUTRE, 3],
-    // [wording.TAUX_OCCUPATION_SEANCES, 4],
+    [wording.TAUX_OCCUPATION_SEANCES, 4],
   ])('ferme l’info bulle après avoir cliqué sur le bouton "Fermer" (%s)', (titreSection, identifiant) => {
     // GIVEN
     renderFakeComponent(<SessionProvider session={mockSession}><PageÉtablissementTerritorialMédicoSocial rechercheViewModel={rechercheViewModel} établissementTerritorialViewModel={établissementTerritorialMédicoSocial} /> </SessionProvider>);