+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog d'Helios
 
+## [1.7.5](https://github.com/DNUM-SocialGouv/helios/compare/release/1.7.4...release/1.7.5) (TODO)
+
+### Bug Fixes
+
+- hel-947: Les libellés des profils sont maintenant obligatoire
+- hel-929: Non affichage du volet d’autorisation
+
+### Features
+
+- hel-919: Ajout de l'indicateur de taux d'occupation global
+
 ## [1.7.4](https://github.com/DNUM-SocialGouv/helios/compare/release/1.7.3...release/1.7.4) (2025-12-17)
 
 ### Bug Fixes